moved chat and message display to their own components
